+++ R/robust_estimation.R
@@ -2,7 +2,7 @@
 #' @export
 standard_error_robust <- function(model,
                                   vcov_estimation = "HC",
-                                  vcov_type = c("HC3", "const", "HC", "HC0", "HC1", "HC2", "HC4", "HC4m", "HC5"),
+                                  vcov_type = c("HC3", "const", "HC", "HC0", "HC1", "HC2", "HC4", "HC4m", "HC5", "CR0", "CR1", "CR1p", "CR1S", "CR2", "CR3"),
                                   vcov_args = NULL,
                                   ...) {
 
@@ -30,7 +30,7 @@ standard_error_robust <- function(model,
 #' @export
 p_value_robust <- function(model,
                            vcov_estimation = "HC",
-                           vcov_type = c("HC3", "const", "HC", "HC0", "HC1", "HC2", "HC4", "HC4m", "HC5"),
+                           vcov_type = c("HC3", "const", "HC", "HC0", "HC1", "HC2", "HC4", "HC4m", "HC5", "CR0", "CR1", "CR1p", "CR1S", "CR2", "CR3"),
                            vcov_args = NULL,
                            ...) {
   # match arguments
@@ -54,7 +54,7 @@ p_value_robust <- function(model,
 ci_robust <- function(model,
                       ci = 0.95,
                       vcov_estimation = "HC",
-                      vcov_type = c("HC3", "const", "HC", "HC0", "HC1", "HC2", "HC4", "HC4m", "HC5"),
+                      vcov_type = c("HC3", "const", "HC", "HC0", "HC1", "HC2", "HC4", "HC4m", "HC5", "CR0", "CR1", "CR1p", "CR1S", "CR2", "CR3"),
                       vcov_args = NULL,
                       ...) {
   vcov_type <- match.arg(vcov_type)
@@ -78,22 +78,36 @@ ci_robust <- function(model,
 #' @importFrom stats coef pnorm pt
 .robust_covariance_matrix <- function(x,
                        vcov_fun = "vcovHC",
-                       vcov_type = c("HC3", "const", "HC", "HC0", "HC1", "HC2", "HC4", "HC4m", "HC5"),
+                       vcov_type = c("HC3", "const", "HC", "HC0", "HC1", "HC2", "HC4", "HC4m", "HC5", "CR0", "CR1", "CR1p", "CR1S", "CR2", "CR3"),
                        vcov_args = NULL) {
-
-  if (!requireNamespace("sandwich", quietly = TRUE)) {
-    stop("Package `sandwich` needed for this function. Please install and try again.")
-  }
 
   # match arguments
   vcov_type <- match.arg(vcov_type)
+
+  # check if required package is available
+  if (vcov_type %in% c("CR0", "CR1", "CR1p", "CR1S", "CR2", "CR3")) {
+    if (!requireNamespace("clubSandwich", quietly = TRUE)) {
+      stop("Package `clubSandwich` needed for this function. Please install and try again.")
+    }
+    package <- "clubSandwich"
+  } else {
+    if (!requireNamespace("sandwich", quietly = TRUE)) {
+      stop("Package `sandwich` needed for this function. Please install and try again.")
+    }
+    package <- "sandwich"
+  }
 
   # get coefficients
   est <- stats::coef(x)
 
   # compute robust standard errors based on vcov
-  vcov_fun <- get(vcov_fun, asNamespace("sandwich"))
-  .vcov <- do.call(vcov_fun, c(list(x = x, type = vcov_type), vcov_args))
+  if (package == "sandwich") {
+    vcov_fun <- get(vcov_fun, asNamespace("sandwich"))
+    .vcov <- do.call(vcov_fun, c(list(x = x, type = vcov_type), vcov_args))
+  } else {
+    vcov_fun <- clubSandwich::vcovCR
+    .vcov <- do.call(vcov_fun, c(list(obj = x, type = vcov_type), vcov_args))
+  }
 
   se <- sqrt(diag(.vcov))
   dendf <- degrees_of_freedom(x, method = "any")