+++ NEWS.md
@@ -28,6 +28,7 @@
 
 ## Changes to functions
 
+- Robust estimation (like `standard_error_robust()` or `ci_robust()`) can now also compute cluster-robust variance-covariance matrices, using the *clubSandwich*  package.
 - `model_parameters()` gets a `robust`-argument, to compute robust standard errors, and confidence intervals and p-values based on robust standard errors.
 - Arguments `p_method` and `ci_method` in `model_parameters.merMod()` were replaced by a single argument `df_method`.
 - `model_parameters.principal()` includes a `MSA` column for objects from `principal_components()`.